requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite
colorama
cachetools
python-dateutil
python-box
colorthief
emoji
dhooks
natural
parsedatetime
motor
sanic
dnspython